codetext control that reuses the same editor as html editor via lang prop
