+++ apps/builder/app/builder/shared/code-editor.tsx
@@ -1,386 +1,122 @@
+import { forwardRef, useMemo, type ComponentProps, useEffect } from "react";
 import {
-  useEffect,
-  useRef,
-  type ReactNode,
-  useState,
-  forwardRef,
-  type ComponentProps,
-} from "react";
-import {
-  Annotation,
-  EditorState,
-  StateEffect,
-  type Extension,
-} from "@codemirror/state";
-import {
-  EditorView,
-  drawSelection,
-  dropCursor,
   keymap,
+  tooltips,
+  highlightSpecialChars,
+  highlightActiveLine,
 } from "@codemirror/view";
+import { bracketMatching, indentOnInput } from "@codemirror/language";
 import {
-  defaultKeymap,
-  history,
-  historyKeymap,
-  indentWithTab,
-} from "@codemirror/commands";
-import { HighlightStyle, syntaxHighlighting } from "@codemirror/language";
-import { tags } from "@lezer/highlight";
-import {
-  theme,
-  textVariants,
-  css,
-  SmallIconButton,
-  Dialog,
-  DialogTrigger,
-  DialogContent,
-  Grid,
-  DialogTitle,
-  Button,
-  DialogClose,
-  Flex,
-} from "@webstudio-is/design-system";
-import { CrossIcon, MaximizeIcon, MinimizeIcon } from "@webstudio-is/icons";
+  autocompletion,
+  closeBrackets,
+  closeBracketsKeymap,
+  completionKeymap,
+} from "@codemirror/autocomplete";
+import { html } from "@codemirror/lang-html";
+import { theme, textVariants, css } from "@webstudio-is/design-system";
+import { BaseCodeEditor, getMinMaxHeightVars } from "./base-code-editor";
 
-const ExternalChange = Annotation.define<boolean>();
-
-const minHeightVar = "--ws-code-editor-min-height";
-const maxHeightVar = "--ws-code-editor-max-height";
-
-export const getMinMaxHeightVars = ({
-  minHeight,
-  maxHeight,
-}: {
-  minHeight: string;
-  maxHeight: string;
-}) => ({
-  [minHeightVar]: minHeight,
-  [maxHeightVar]: maxHeight,
-});
-
-const editorContentStyle = css({
-  ...textVariants.mono,
-  // fit editor into parent if stretched
-  display: "flex",
-  minHeight: 0,
-  boxSizing: "border-box",
-  color: theme.colors.foregroundMain,
-  borderRadius: theme.borderRadius[4],
-  border: `1px solid ${theme.colors.borderMain}`,
-  background: theme.colors.backgroundControls,
-  paddingTop: 6,
-  paddingBottom: 4,
-  paddingRight: theme.spacing[2],
-  paddingLeft: theme.spacing[3],
-  "&:focus-within": {
-    borderColor: theme.colors.borderFocus,
-    outline: `1px solid ${theme.colors.borderFocus}`,
-  },
-  '&[data-invalid="true"]': {
-    borderColor: theme.colors.borderDestructiveMain,
-    outlineColor: theme.colors.borderDestructiveMain,
-  },
-  "& .cm-focused": {
-    outline: "none",
-  },
-  // fix scrolls appear on mount
-  "& .cm-scroller": {
-    overflowX: "hidden",
-  },
-  "& .cm-content": {
-    padding: 0,
-    // makes sure you can click to focus when editor content is smaller than the container
-    minHeight: "100%",
-  },
-  "& .cm-line": {
-    padding: 0,
-  },
-  "& .cm-editor": {
-    width: "100%",
-    // avoid modifying height in .cm-content
-    // because it breaks scroll events and makes scrolling laggy
-    minHeight: `var(${minHeightVar}, auto)`,
-    maxHeight: `var(${maxHeightVar}, none)`,
+const autocompletionStyle = css({
+  "&.cm-tooltip.cm-tooltip-autocomplete": {
+    ...textVariants.mono,
+    border: "none",
+    backgroundColor: "transparent",
+    // override none set on body by radix popover
+    pointerEvents: "auto",
+    "& ul": {
+      minWidth: 160,
+      maxWidth: 260,
+      width: "max-content",
+      boxSizing: "border-box",
+      borderRadius: theme.borderRadius[6],
+      backgroundColor: theme.colors.backgroundMenu,
+      border: `1px solid ${theme.colors.borderMain}`,
+      boxShadow: `${theme.shadows.menuDropShadow}, inset 0 0 0 1px ${theme.colors.borderMenuInner}`,
+      padding: theme.spacing[3],
+      "& li": {
+        ...textVariants.labelsTitleCase,
+        textTransform: "none",
+        position: "relative",
+        display: "flex",
+        alignItems: "center",
+        color: theme.colors.foregroundMain,
+        padding: theme.spacing[3],
+        borderRadius: theme.borderRadius[3],
+        "&[aria-selected], &:hover": {
+          color: theme.colors.foregroundMain,
+          backgroundColor: theme.colors.backgroundItemMenuItemHover,
+        },
+        "& .cm-completionLabel": {
+          flexGrow: 1,
+        },
+        "& .cm-completionDetail": {
+          overflow: "hidden",
+          textOverflow: "ellipsis",
+          fontStyle: "normal",
+          color: theme.colors.hint,
+        },
+      },
+    },
   },
 });
 
-// https://thememirror.net/clouds
-const highlightStyle = HighlightStyle.define([
-  {
-    tag: tags.comment,
-    color: "#BCC8BA",
-  },
-  {
-    tag: [tags.string, tags.special(tags.brace), tags.regexp],
-    color: "#5D90CD",
-  },
-  {
-    tag: [tags.number, tags.bool, tags.null],
-    color: "#46A609",
-  },
-  {
-    tag: tags.keyword,
-    color: "#AF956F",
-  },
-  {
-    tag: [tags.definitionKeyword, tags.modifier],
-    color: "#C52727",
-  },
-  {
-    tag: [tags.angleBracket, tags.tagName, tags.attributeName],
-    color: "#606060",
-  },
-  {
-    tag: tags.self,
-    color: "#000",
-  },
-]);
-
-type EditorContentProps = {
-  extensions: Extension[];
-  readOnly?: boolean;
-  autoFocus?: boolean;
-  invalid?: boolean;
-  value: string;
-  onChange: (newValue: string) => void;
-  onBlur?: (event: FocusEvent) => void;
-};
-
-export const BaseCodeEditor = ({
-  extensions,
-  readOnly = false,
-  autoFocus = false,
-  invalid = false,
-  value,
-  onChange,
-  onBlur,
-}: EditorContentProps) => {
-  const editorRef = useRef<null | HTMLDivElement>(null);
-  const viewRef = useRef<undefined | EditorView>();
-
-  const onChangeRef = useRef(onChange);
-  onChangeRef.current = onChange;
-  const onBlurRef = useRef(onBlur);
-  onBlurRef.current = onBlur;
-
-  // setup editor
-
-  useEffect(() => {
-    if (editorRef.current === null) {
-      return;
-    }
-    const view = new EditorView({
-      doc: "",
-      parent: editorRef.current,
-    });
-    if (autoFocus) {
-      view.focus();
-    }
-    viewRef.current = view;
-    return () => {
-      view.destroy();
-    };
-  }, [autoFocus]);
-
-  // update extensions whenever variables data is changed
-
-  useEffect(() => {
-    const view = viewRef.current;
-    if (view === undefined) {
-      return;
-    }
-    view.dispatch({
-      effects: StateEffect.reconfigure.of([
-        ...extensions,
-        history(),
-        drawSelection(),
-        dropCursor(),
-        syntaxHighlighting(highlightStyle, { fallback: true }),
-        keymap.of([...defaultKeymap, ...historyKeymap, indentWithTab]),
-        EditorView.lineWrapping,
-        EditorView.editable.of(readOnly === false),
-        EditorState.readOnly.of(readOnly === true),
-        // https://github.com/uiwjs/react-codemirror/blob/5d7a37245ce70e61f215b77dc42a7eaf295c46e7/core/src/useCodeMirror.ts#L57-L70
-        EditorView.updateListener.of((update) => {
-          if (
-            // prevent invoking callback when focus or selection is changed
-            update.docChanged &&
-            // prevent invoking callback when the change came from react value
-            update.transactions.some((trx) =>
-              trx.annotation(ExternalChange)
-            ) === false
-          ) {
-            onChangeRef.current(update.state.doc.toString());
-          }
-        }),
-        EditorView.domEventHandlers({
-          blur(event) {
-            onBlurRef.current?.(event);
-          },
-          cut(event) {
-            // prevent catching cut by global copy paste
-            // with target outside of contenteditable
-            event.stopPropagation();
-          },
-        }),
-      ]),
-    });
-  }, [readOnly, extensions]);
-
-  // update editor with react value
-  // https://github.com/uiwjs/react-codemirror/blob/5d7a37245ce70e61f215b77dc42a7eaf295c46e7/core/src/useCodeMirror.ts#L158-L169
-  useEffect(() => {
-    const view = viewRef.current;
-    if (view === undefined) {
-      return;
-    }
-    // prevent updating when editor has the same state
-    // and can be the source of new value
-    if (value === view.state.doc.toString()) {
-      return;
-    }
-    view.dispatch({
-      changes: { from: 0, to: view.state.doc.length, insert: value },
-      annotations: [ExternalChange.of(true)],
-    });
-  }, [value]);
-
-  return (
-    <div
-      className={editorContentStyle()}
-      data-invalid={invalid}
-      ref={editorRef}
-    />
-  );
-};
-
-const editorDialogControlStyle = css({
+const wrapperStyle = css({
   position: "relative",
-  "&:hover": {
-    "--ws-code-editor-maximize-icon-visibility": "visible",
-  },
+  // 1 line is 16px
+  // set min 10 lines and max 20 lines
+  ...getMinMaxHeightVars({ minHeight: "160px", maxHeight: "320px" }),
 });
 
-export const EditorDialogControl = ({ children }: { children: ReactNode }) => {
-  return <div className={editorDialogControlStyle()}>{children}</div>;
-};
+const getHtmlExtensions = () => [
+  highlightActiveLine(),
+  highlightSpecialChars(),
+  indentOnInput(),
+  html({}),
+  bracketMatching(),
+  closeBrackets(),
+  // render autocomplete in body
+  // to prevent popover scroll overflow
+  tooltips({ parent: document.body }),
+  autocompletion({
+    icons: false,
+    tooltipClass: () => autocompletionStyle.toString(),
+    closeOnBlur: false,
+  }),
+  keymap.of([...closeBracketsKeymap, ...completionKeymap]),
+];
 
-export const EditorDialogButton = forwardRef<
-  HTMLButtonElement,
-  Partial<ComponentProps<typeof SmallIconButton>>
->((props, ref) => {
+export const CodeEditor = forwardRef<
+  HTMLDivElement,
+  Omit<ComponentProps<typeof BaseCodeEditor>, "extensions"> & { lang?: "html" }
+>(({ lang, ...props }, ref) => {
+  const extensions = useMemo(
+    () => (lang === "html" ? getHtmlExtensions() : []),
+    []
+  );
+
+  // prevent clicking on autocomplete options propagating to body
+  // and closing dialogs and popovers
+  useEffect(() => {
+    const handlePointerDown = (event: PointerEvent) => {
+      if (
+        event.target instanceof HTMLElement &&
+        event.target.closest(".cm-tooltip-autocomplete")
+      ) {
+        event.stopPropagation();
+      }
+    };
+    const options = { capture: true };
+    document.addEventListener("pointerdown", handlePointerDown, options);
+    return () => {
+      document.removeEventListener("pointerdown", handlePointerDown, options);
+    };
+  }, []);
+
   return (
-    <SmallIconButton
-      {...props}
-      ref={ref}
-      icon={<MaximizeIcon />}
-      css={{
-        position: "absolute",
-        top: 6,
-        right: 4,
-        visibility: `var(--ws-code-editor-maximize-icon-visibility, hidden)`,
-      }}
-    />
+    <div className={wrapperStyle()} ref={ref}>
+      <BaseCodeEditor {...props} extensions={extensions} />
+    </div>
   );
 });
-EditorDialogButton.displayName = "EditorDialogButton";
 
-export const EditorDialog = ({
-  open,
-  onOpenChange,
-  title,
-  content,
-  children,
-}: {
-  open?: boolean;
-  onOpenChange?: (newOpen: boolean) => void;
-  title?: ReactNode;
-  content: ReactNode;
-  children: ReactNode;
-}) => {
-  const [isMaximized, setIsMaximized] = useState(false);
-  return (
-    <Dialog open={open} onOpenChange={onOpenChange} modal={false}>
-      <DialogTrigger asChild>{children}</DialogTrigger>
-      <DialogContent
-        resize="auto"
-        width={640}
-        height={480}
-        minHeight={240}
-        isMaximized={isMaximized}
-        onInteractOutside={(event) => {
-          event.preventDefault();
-        }}
-      >
-        <Grid
-          align="stretch"
-          css={{
-            padding: theme.spacing[7],
-            height: "100%",
-            overflow: "hidden",
-            boxSizing: "content-box",
-          }}
-        >
-          {content}
-        </Grid>
-        {/* Title is at the end intentionally,
-         * to make the close button last in the tab order
-         */}
-        <DialogTitle
-          draggable
-          suffix={
-            <Flex
-              gap="1"
-              onMouseDown={(event) => {
-                // Prevent dragging dialog
-                event.preventDefault();
-              }}
-            >
-              <Button
-                color="ghost"
-                prefix={isMaximized ? <MinimizeIcon /> : <MaximizeIcon />}
-                aria-label="Expand"
-                onClick={() => setIsMaximized(isMaximized ? false : true)}
-              />
-              <DialogClose asChild>
-                <Button
-                  color="ghost"
-                  prefix={<CrossIcon />}
-                  aria-label="Close"
-                />
-              </DialogClose>
-            </Flex>
-          }
-        >
-          {title}
-        </DialogTitle>
-      </DialogContent>
-    </Dialog>
-  );
-};
-
-export const CodeEditor = ({
-  title,
-  open,
-  onOpenChange,
-  ...editorContentProps
-}: EditorContentProps & {
-  title?: ReactNode;
-  open?: boolean;
-  onOpenChange?: (newOpen: boolean) => void;
-}) => {
-  const content = <BaseCodeEditor {...editorContentProps} />;
-  return (
-    <EditorDialogControl>
-      {content}
-      <EditorDialog
-        open={open}
-        onOpenChange={onOpenChange}
-        title={title}
-        content={content}
-      >
-        <EditorDialogButton />
-      </EditorDialog>
-    </EditorDialogControl>
-  );
-};
+CodeEditor.displayName = "CodeEditor";
